requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Include production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.2.0
hypothesis>=6.100.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pylint>=2.17.0
mypy>=1.4.0
mutmut>=2.4.4

# Type stubs
types-aiohttp
pandas-stubs

# Security
bandit>=1.7.5
safety>=2.3.5

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Development tools
ipython>=8.14.0
ipdb>=0.13.13
